Refactors genesis flash cart saves to reflect the source code file structure
